Add files via upload
The internet has been a platform for individuals, groups of people and companies to interact
with one another through the social media. The social media has truly aided interaction and
even other business services through social networks, forums, blogs, etc.
Social media apps have taken up a great part of our lives and the application market already,
and their number continues to grow. With the increase of smartphone usage worldwide there
have emerged the opportunities for social networks with more advanced features.
Social media helps people establish better relationships with their family and friends, and
now the networking sites also show their significance for apps.
That is why people spend a lot of their time online browsing social sites, and usage has only
gone up with increase in the use of smartphones and tablets.
The primary factor that makes social sites such an exciting place for your brand is that your
audience is indeed there. Therefore, interaction with your clients becomes more accessible.
Every second, thousands of people share relevant information about their profile or their
needs and demands digitally, either by a blog comment, on their status, in various ways, and
this can end up in generating insights and adding value to enterprises.
Thus, it is clear that social sites are vital for the development of a sound work plan that
reflects on the acquisition of new customers and consequently increases the number of sales
and downloads.
Companies can use these platforms differently and obtain numerous advantages and benefits.
See the following main advantages of social sites for the expansion of your app.
YouConnect is a social media application with rudimentary features inspired by existing
social media platforms such as Facebook and Twitter. The main purpose of this application is
to demonstrate how to use the MERN stack technologies to implement features that allow
users to connect and interact over content. We can extend these implementations further, as
desired, for more complex features.